+++ navigation/atlassian-switcher/src/admin/utils/admin-link-collector.ts
@@ -1,4 +1,7 @@
-import { getAdministrationLinks } from './admin-links';
+import {
+  getAdministrationLinks,
+  getAdministrationLinksNext,
+} from './admin-links';
 import { isComplete, isError } from '../../common/providers/as-data-provider';
 import { ProviderResults } from '../../types';
 
@@ -16,6 +19,22 @@ export function collectAdminLinks(
       return getAdministrationLinks(managePermission.data, adminUrl);
     }
 
+    return [];
+  }
+}
+
+export function collectAdminLinksNext(
+  managePermission: ProviderResults['managePermission'],
+  adminUrl?: string,
+) {
+  if (isError(managePermission)) {
+    return [];
+  }
+
+  if (isComplete(managePermission)) {
+    if (managePermission.data) {
+      return getAdministrationLinksNext(adminUrl);
+    }
     return [];
   }
 }