+++ navigation/atlassian-switcher/src/ui/components/__tests__/unit/trello-switcher.tsx
@@ -39,6 +39,7 @@ describe('trello-switcher', () => {
         isProductStoreInTrelloJSWFirstEnabled: false,
         isProductStoreInTrelloConfluenceFirstEnabled: false,
         isSlackDiscoveryEnabled: false,
+        isTrustedAdminUIDeprecationEnabled: false,
       },
       triggerXFlow: () => {},
       onDiscoverMoreClicked: () => {},