+++ people-and-teams/profilecard/examples/02-profilecard-resourced.tsx
@@ -1,6 +1,8 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import styled from 'styled-components';
+
+import Button, { ButtonGroup } from '@atlaskit/button';
 
 import ProfileCardResourced from '../src';
 
@@ -32,19 +34,66 @@ const defaultProps = {
 };
 
 export default function Example() {
+  const [actions, setActions] = useState(defaultProps.actions);
+  const [resourceClient, setResourceClient] = useState(
+    defaultProps.resourceClient,
+  );
+
+  const handleChangeActions = () => {
+    setActions([
+      {
+        label: 'View profile - ' + Date.now(),
+        id: 'view-profile - ' + Date.now(),
+        callback: () => {},
+      },
+    ]);
+  };
+
+  const handleChangeResourceClient = () => {
+    setResourceClient(getMockProfileClient(10, 0));
+  };
+
   return (
     <LocaleIntlProvider>
       <MainStage>
         <Wrap>
-          <ProfileCardResourced {...defaultProps} userId="1" />
+          <ButtonGroup>
+            {/* Check re-rederning Profile Card when `actions` is changed */}
+            <Button onClick={handleChangeActions}>
+              Update "actions" props
+            </Button>
+            {/* Check re-fetching data when `resourceClient` is changed */}
+            <Button onClick={handleChangeResourceClient}>
+              Update "resourceClient" props
+            </Button>
+          </ButtonGroup>
+        </Wrap>
+
+        <Wrap>
+          <ProfileCardResourced
+            {...defaultProps}
+            userId="1"
+            actions={actions}
+            resourceClient={resourceClient}
+          />
         </Wrap>
         <br />
         <Wrap>
-          <ProfileCardResourced {...defaultProps} userId="2" />
+          <ProfileCardResourced
+            {...defaultProps}
+            userId="2"
+            actions={actions}
+            resourceClient={resourceClient}
+          />
         </Wrap>
         <br />
         <Wrap>
-          <ProfileCardResourced {...defaultProps} userId="error:NotFound" />
+          <ProfileCardResourced
+            {...defaultProps}
+            userId="error:NotFound"
+            actions={actions}
+            resourceClient={resourceClient}
+          />
         </Wrap>
       </MainStage>
     </LocaleIntlProvider>